exam/requirements.txt
asgiref==3.8.1
Django==5.1.3
django-environ==0.11.2
django-grappelli==4.0.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
pillow==11.0.0
PyJWT==2.9.0
sqlparse==0.5.1
